dept_8082模块中配置mybatis,配置映射文件

+++ spring_cloud_learn/provider_dept_8082/src/main/java/com/gsau/provider_dept_8082/dao/DepartmentDao.java
@@ -25,7 +25,7 @@ public interface DepartmentDao {
      * @param deptNo
      * @return
      */
-    Department findById(Long deptNo);
+    Department findById(int deptNo);
 
     /**
      * 查询全部
